Add Filter type
