--- src/components/FilterButtons.module.css
@@ -1,36 +1,0 @@
-.filters {
-  display: flex;
-  gap: 0.75rem;
-}
-
-.button {
-  padding: 0.5rem 1rem;
-  background: #e5e7eb;
-  border: none;
-  border-radius: 12px;
-  cursor: pointer;
-  transition: background-color 0.2s ease, transform 0.2s ease;
-}
-
-.button:hover {
-  background: #d1d5db;
-  transform: scale(1.05);
-}
-
-.activeTodos {
-  background: #fb923c;
-  color: white;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.08);
-}
-
-.activeProximos {
-  background: #22c55e;
-  color: white;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.08);
-}
-
-.activePassados {
-  background: #3b82f6;
-  color: white;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.08);
-}
