+++ src/components/StatsCards.module.css
@@ -11,6 +11,12 @@
   border-radius: 12px;
   border-left: 6px solid transparent;
   box-shadow: 0 1px 3px rgba(0,0,0,0.08);
+  cursor: pointer;
+  transition: background-color 0.2s ease, transform 0.2s ease;
+}
+
+.card:hover {
+  transform: scale(1.02);
 }
 
 .total {
@@ -23,4 +29,19 @@
 
 .passado {
   border-color: #6b7280;
+}
+
+.activeTotal {
+  background: #fb923c;
+  color: white;
+}
+
+.activeFuturo {
+  background: #22c55e;
+  color: white;
+}
+
+.activePassado {
+  background: #6b7280;
+  color: white;
 }
